+++ lib/ui/gradient_container.dart
@@ -1,0 +1,20 @@
+import 'package:first_app/ui/styled_text.dart';
+import 'package:flutter/material.dart';
+
+class GradientContainer extends StatelessWidget {
+  const GradientContainer({super.key});
+
+  @override
+  Widget build(context) {
+    return Container(
+      decoration: const BoxDecoration(
+        gradient: LinearGradient(colors: [
+          Colors.white,
+          Color.fromARGB(255, 235, 195, 242),
+          Color.fromARGB(255, 160, 74, 175)
+        ], begin: Alignment.topLeft, end: Alignment.bottomRight),
+      ),
+      child: const Center(child: StyledText()),
+    );
+  }
+}
